+++ src/main/java/com/nnxx/service/IVisaOfficersService.java
@@ -20,7 +20,7 @@ public interface IVisaOfficersService extends IService<VisaOfficers> {
 
     Result delete(Integer id);
 
-    Result selectOne(Integer id);
+    Result selectOne(Long id);
 
     Result selectAll();
