+++ src/main/java/com/nnxx/service/impl/StudyAdvisorsServiceImpl.java
@@ -49,7 +49,7 @@ public class StudyAdvisorsServiceImpl extends ServiceImpl<StudyAdvisorsMapper, S
     }
 
     @Override
-    public Result selectOne(Integer id) {
+    public Result selectOne(Long id) {
         //用户自己查询留学顾问
         StudyAdvisors studyAdvisors = query().eq("user_id",id).one();
         int status = studyAdvisors!=null? Code.SELECT_YES:Code.SELECT_ERROR;
